box models: implement time step ramp up
the biggest open issue is how to combine this with partial jacobian
reassembly. (it works as it is implemented now but the percentage of
elements which needs to be reassembled is quite high between the
newton iterations.) one idea is to just reassemble a fixed fraction of
vertices, e.g. 20%. Here the problem is how to determine this set of
vertices efficiently...

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@4066 2fb0f335-1f38-0410-981e-8018bf24f1b0