box models: implement hints
using hints may dramatically speed up the expensive computations of
the volume variables if the calculations are expensive. This is due to
the fact that the initial guess for non-linear computation is usually
much closer to the result when using hints. (this mainly affects
calculation of the thermodynamic equilibrium when using non-linear
fluid systems.)

hints must be explicitly enabled by setting the "EnableHints" property
to true.

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@5456 2fb0f335-1f38-0410-981e-8018bf24f1b0