box models: make solution dependent neumann conditions and the source term non-disruptive
now the problems do not need to specify 'using ParentType::neumann'
anymore. this is pretty much a temporary API, the real solution would
be that the methods get a context plus an entity index and an entity
codimension. the context stores the current solution and volume
variables, as well as the current element and its element
geometry. for other discretizations it may store other things...

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@5462 2fb0f335-1f38-0410-981e-8018bf24f1b0