box models: introduce method updatePVWeights()
allows to update the relative weights of the primary variables based
on the values of the volume variables. this makes it possible to use
expensive criteria as weighting factors of the primary
variables. Currently, the only "real" user is the Mp-Nc model which
now defines the weight of the component fugacity as the minimum of the
fugacity coefficient times the phase pressure of any phase. the 2p2c
model would possibly benefit from a similar treatment, though...

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@6413 2fb0f335-1f38-0410-981e-8018bf24f1b0