introducing quad precision floating point support
this is an aid for debugging precision problems. in order to use this,
add --enable-quad to the CONFIGURE_FLAGS of your options file, re-run
dunecontrol, include the header file "dumux/common/quad.hh" after
config.h and *before* everything else. next, set the Scalar property
to 'quad'. finally, you need to patch Dune::className() to not access
the typeid method, for example by returning a constant string like
"<unknown>" (so far, typeids are not yet generated for quad precision
floating point values by GCC which I consider to be a compiler bug...)

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@6534 2fb0f335-1f38-0410-981e-8018bf24f1b0