property system: deprecate property defaults, make the PTAG() optional in GET_PROP*()
defaults for properties come with quite a few issues, probably the
most galling one is that they can only be specified once in the whole
code base. This means so they are best avoided.

Further, the PTAG() in the GET_PROP* is not useful in 99% of all
cases. It was there to also allow full c++ types (e.g. template
arguments) as parameters. For the code that really needs this, there
now is GET_PROP*_ and PTAG_ (mind the underscores). To my knowledge
that only affects the run-time parameter system. The parameter system
has been adapted, everyone else can now write GET_PROP_TYPE(TypeTag,
Scalar) instead of GET_TYPE_PROP(TypeTag, PTAG(Scalar)). (the PTAG
macro is still around but just does nothing.)

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@7072 2fb0f335-1f38-0410-981e-8018bf24f1b0