fix valgrind complaint in the h2o-n2 fluid system
now we assume an ideal mixture of the gas phase again. we probably
want that, as the 2p2c model as currently implemented does not support
non-ideal mixtures, because the MiscibleMultiPhaseComposition
constraint solver does not support this yet. (On a different note, a
reference is missing for the approach taken for the fugacity
coefficients of the gas phase for the "gas as non-ideal mixture"
code.)

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@7096 2fb0f335-1f38-0410-981e-8018bf24f1b0