partial reassemble: only update the solution if the DOF gets reassembled
also make the MpNc model respect absolute convergence criteria. IMHO
an unweighted absolute criterion is pretty useless in most contexts,
though. that's because the criterion is about the amount of
conservation quantities that get lost per second but a losing or
gaining a joule of energy is much less of a problem than losing or
gaining a kilogram of mass...

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@7442 2fb0f335-1f38-0410-981e-8018bf24f1b0