more interesting and more reasonable test-problems for the 1p and 1p2c model
-1ptest_problem is now a lens problem, the  lens can be defined in the  parameter
file, documentation is corrected

-1p2ctest is not a bio-problem anymore, but uses a new 1phase
fluidsystem for component transport of N2 in water. The problem includes
an outflow condition, the documentation is corrected (for example:
compressible fluids are also possible)

-1p2cmodel: use phaseIdx (set in 1p2cindices) instead of hard coded
index 0 for the phase, so that it is in general possible to use i.e. a
gas phase (see stokes2c as example)


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@7686 2fb0f335-1f38-0410-981e-8018bf24f1b0