decoupled: store the grid view by value not by reference in the variable class
this is a common mistake which also used to be present in the box
models and in dune-multidomaingrid: Grid views should never be stored
by reference as the object may vanish as soon as it gets out of scope
in the calling function. given the fact that grid views in dune are
created on the fly, this leads to subtle bugs.

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@7902 2fb0f335-1f38-0410-981e-8018bf24f1b0