Changed component indices: phaseIdx is called from the property system (default to nPhase)
wComp or comp1Idx -> transportCompIdx (transported component)
nComp or comp0Idx -> phaseCompIdx (main component of phase)

transportCompIdx is |1-phaseIdx|, because usually the transported
component of interest is the minor component of the phase. This allows
the to choose an arbitrary phase of a two-phase fluidsystem, if
wComp is the main component of wPhase and both have the same index with either 0 or 1.

Corrected index names for primary variables, they get the correct
primary variable index instead of the equation index;
massOrMoleFracIdx has been employed for the mass or mole fraction (seems
to be self-explanatory...)

Adapted also test problems for stokes2c, stokes2cni.

Reviewed bei Markus.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@8275 2fb0f335-1f38-0410-981e-8018bf24f1b0