the default of the beta scaling parameters did not work. Now fixed.
What are the beta scaling parameters?

They allow a scaling of two-phase capillary pressures according to
the surface tensions. Typically, the Gas-Water system is the reference
and the betaGW parameter is 1. Surface tensions between the other
fluid phases are a little different, and this can be taken into
account phenomenologically by the beta scaling parameters. An example
is given in dumux-lecture/mm/columnxylene exercise.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@8424 2fb0f335-1f38-0410-981e-8018bf24f1b0