Minor changes -remove disabled acess functions in multiphysics cell data -enable "EnableCapillarity" Property for Multiphyiscs in simple update function -avoid wrong output: 	* set error correction to 0 if it is 0 	* upwind information in simple subdomain set for both phases 	* First episode has index 1 and not 0 => fix first timestep size in fvpressurecompositional
Fixed bug for Adaptive Multiphysics (in devel):
After grid modification, a once complex local cellData object might now point to a cell in simple subdomain. Hence both mobilities have to be overwritten by material law update, which was not done in simple update. This only has an effect when grid is adapted.
reviewed by Philipp

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@9501 2fb0f335-1f38-0410-981e-8018bf24f1b0