dumux/dumux/implicit/common/implicitmodel dumux/dumux/nonlinear/newtoncontroller
reintroducing some function for checking whether the current solution makes sense: checkPlausibility().
After each newton iteration the newtoncontroller asks the model to check. 
So far mainly used for debugging: the spatial coordinate of the bad value is printed to screen and an exception is thrown. 
See mpncmodelkinetic for a possible implementation

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@11057 2fb0f335-1f38-0410-981e-8018bf24f1b0