make calculation of gas mixture enthalpy consistent with the other fluidsystems: call component enthalpies with phase pressure not with partial pressure
this was discussed with Klaus, Holger, Philipp and Bernd. The
enthalpies can be called with the partial pressure and be summed up directly,
or they can be called  with the phase pressure and are then weighted by
the mass fractions. Klaus compared the two possibilities and found only
minor differences.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@12402 2fb0f335-1f38-0410-981e-8018bf24f1b0