Ignore last commit ;-)
AMG is used as the primary solver to allow parallel computations. If neither AMG nor PDELab are available, SuperLU will be used, which was the solver previously implemented. If neither of these solvers are available, there will be a warning.

In addition, the grid was changed from SGrid to YaspGrid.

Reviewed by alexk.

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@12849 2fb0f335-1f38-0410-981e-8018bf24f1b0