Use grid view attached to problem and moved initialization to a later stage
In general the gridView is only available from the problem directly. Therefore
this patch does problem.model().gridView()->problem.gridView() to fix
this.

In addition the model might be null during the constructor, which
might be called somewhere in NewtonConroller. It seems like the model
is initialized at a later stage. Therefore we now use a new method
initGhostsAndOwners to populate the arrays needed for the projections
in the parallel solvers. Using a boolean flag we make sure that this
is done only once.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/branches/strip-pdelab@13039 2fb0f335-1f38-0410-981e-8018bf24f1b0