[2p2c] revert commits 13192-3
For the box porous media spatial parameters, we allow 
- both cell-wise and vertex-wise defined permeabilities,
- both scalar and tensorial permeabilities.

The reverted commits assumed vertex-wise scalar permeabilities 
and are not general enough. The currently correct way to give 
the permeabilities to the VTK output is to overwrite the method 
addOutputVtkFields in the problem at hand. There, it should be 
clear what type of permeabilities are used. See 
test/implicit/co2/heterogeneousproblem.hh
for details. Admittedly, this is not a very nice way, a better, but
general solution would be great.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@13197 2fb0f335-1f38-0410-981e-8018bf24f1b0