Make correct use of wildcards.
The correct usage of wildcards is "$(wildcard *.hh)".
With this patch we use these instead of "*.hh".

Quoting from https://www.gnu.org/software/make/manual/html_node/Wildcard-Examples.html#Wildcard-Examples

Wildcard expansion does not happen when you define a variable. Thus, if you write this:

     objects = *.o

then the value of the variable objects is the actual string ‘*.o’. However, if you use the value of objects in a target or prerequisite, wildcard expansion will take place there. If you use the value of objects in a recipe, the shell may perform wildcard expansion when the recipe runs. To set objects to the expansion, instead use:

     objects := $(wildcard *.o)



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/branches/strip-pdelab@13248 2fb0f335-1f38-0410-981e-8018bf24f1b0