[stokes] add possibility to use molar density in continuity equation
For some situations, it appears to be useful to use the molar instead of
the mass density in the Stokes continuity equation. In order to enable
this behavior, the property UseMoles is used for Stokes and the
VolumeVariables are equipped with the required functionality.
Corresponding parts can be deleted from the derived Stokesnc model.

Modified from a patch provided by Vishal, thanks.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@13258 2fb0f335-1f38-0410-981e-8018bf24f1b0