[nonisothermal] introduce niproperties.hh, adapt effective thermal conductivities, correct comments and spacing.
The additional properties required by the generic non-isothermal model
can be centralized in a file niproperties.hh. 

The signature of the effectiveThermalConductivity functions is unified.
They get the volume variables, spatial parameters and the arguments that
will be required for calling the spatial parameters. When attempting to
use the nonisothermal model for 3p3c, it will be seen whether this
signature is genral enough.

Reviewed by Alex K.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@13289 2fb0f335-1f38-0410-981e-8018bf24f1b0