[nonisothermal] move the already existing non-isothermal two-phase models to the generic setting
In order to use a new nonisothermal model, only the includes in
the corresponding application files have to be changed from
including the dedicated non-isothermal model such as
#include <dumux/implicit/2p2cni/2p2cnimodel.hh>
to the corresponding isothermal one such as
#include <dumux/implicit/2p2c/2p2cmodel.hh>
This has been done for the tests in test/implicit/2pni, ...2p2cni and
...co2ni.

For enabling the isothermal models to get the new non-isothermal
functionality, it was sufficient to modify the files ...properties.hh
and ...propertydefaults.hh:

- In ...properties.hh, a non-isothermal type tag such as "TwoPTwoCNI" is
  introduced that inhereits from the corresponding isothermal one AND
  the type tag "NonIsothermal". In order to not mix it up with the old
  non-isothermal models, corresponding safety guards were added.

- In ...propertydefaults.hh, the properties introduced in
  niproperties.hh of the isothermal model have to be set.

The old non-isothermal models have been deprecated by emitting a warning
if the corresponding ...properties.hh file is included. To ensure that
they still work, the ...properties.hh file is included on top of every
file that corresponds to the old model. Together with the safety guard
mentioned above, this guarantees that in fact the old model is chosen if
desired.

Reviewed by Alex K.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@13290 2fb0f335-1f38-0410-981e-8018bf24f1b0