[nonisothermal] improve naming and meaning of solid thermal parameters
This implements FS#216.

The "heatCapacity" function of the spatial parameters and the volume
variables for the implicit nonisothermal models was a misnomer, since it
returned an effective quantity, namely, 
heatCapacity*density*(1 - porosity) in [J/(K m^3)].
Except for mpnc, which resulted in an additional inconsistency.

Corresponding to the decision documented in FS#216, this patch renames
the function to "solidHeatCapacity" and returns always the "true"
(non-effective) heat capacity in [J/(kg K)]. This requires an additional
function "solidDensity" which returns the mass density of the porous
matrix. Moreover, the functions "thermalConductivitySolid/Fluid" are
renamed to "solid/fluidThermalConductivity". The decision to prepend
with "solid/fluid" rather than to append is motivated by consistency
with components and fluid systems, where "gas" and "liquid" are always
prepended to the corresponding function names.

_Beware_: this change breaks compatibility. You have to adapt your
spatial parameters such that they offer functions "solidHeatCapacity",
"solidDensity" and "solidThermalConductivity".

Reviewed by Alex.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14070 2fb0f335-1f38-0410-981e-8018bf24f1b0