[newton] improve terminology
The terminology for the Newton method is improved according to FS#238.
In particular, what has been referred to as "relative error" is now
termed "maximum relative shift", while "absolute error" has been renamed
to "residual reduction". This is particularly important, if
corresponding parameters or properties are set. See the CHANGELOG for
details.

Reviewed by gruenich.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14132 2fb0f335-1f38-0410-981e-8018bf24f1b0