[newton] increase backwards compatibility for the recent change in terminology
1. The deprecated protected member variables now are references to the
new variables. This ensures that the right value is considered even if
the old name is used.

2. In the constructor of the NewtonController, a check is performed if
a value for an old parameter name is set run-time. If yes, this value is
used correspondingly. A warning is still printed when the controller is
destructed.

This should ensure that the same results are produced even if the old
names are used.

Reviewed by gruenich.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14140 2fb0f335-1f38-0410-981e-8018bf24f1b0