[el2p] parallelize the el2p model
So far, the el2p model did not run in parallel, since our parallel
solvers assume a certain ordering of the degrees of freedom. This patch
mainly introduces an AMG backend that is tailored to the el2p model. It
simply reorders the linear system by copying from the original system,
passes the reordered system to the original AMG backend, and
reorders/copies the result back. This could be achieved more
efficiently and with less memory usage by already reordering the dofs
inside the el2p model or by adapting the AMG backend. Since the
performance impact appears to be minimal, we decided to keep the current
simple solution and try a better one once we need it.

Reviewed by Martin B.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14235 2fb0f335-1f38-0410-981e-8018bf24f1b0