[implicit] store Geometry if it is re-used
Since grid implementations like Yasp calculate the Geometries on the
fly, every call to entity.geometry() may involve computation. This patch
stores a Geometry when it can be used later on in the same block. Only
for the implicit models, decoupled contains a lot more.

For simple models this actually pays off: for 1p in 3d on a larger grid,
the running time could be reduced from 15.5 to 14 s.

Reviewed by Timo.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14284 2fb0f335-1f38-0410-981e-8018bf24f1b0