[newton] fix line search
This fixes a bug that resulted from the recent changes to the Newton
convergence measures. Instead of storing the residual, the residual
reduction is stored, but it has been forgotten to initialize this
reduction with 1.0 instead of the initial residual.

This only affected the line search where the current residual/reduction
is compared to the last one.

Reviewed by Alex.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@14895 2fb0f335-1f38-0410-981e-8018bf24f1b0