[boundary conditions] setDirichlet: improve documentation and usage
The general setDirichlet function takes two parameters (pvIdx, eqIdx)
corresponding to the index of the primary variable and the index of the
balance equation. A convenience function that only takes one parameter
exists which calls the general version with two equal parameters. So
far, this one parameter has been documented as equation index. However,
it is better to refer to the parameter as a primary variable index.

This commit changes the parameter name and improves the documentation.
Furthermore, it replaces calls in the tests, whenever the general
version was called with two parameters of different name but equal value
or whenever the simplified version was called with an equation index
instead of a primary variable index.

See FS#270 for details.

Reviewed by Timo.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@15115 2fb0f335-1f38-0410-981e-8018bf24f1b0