[implicit] add a test that uses the general setDirichlet method
The new test employs the function
setDirichlet(pvIdx, eqIdx)
with pvIdx != eqIdx.

On the right boundary, a Dirichlet value for the nonwetting-phase
pressure should be set. At the same time, nonwetting phase should
be allowed to leave the domain by means of an outflow boundary
condition. In order to achieve this the general 'setDirichlet'
method is used in the function 'boundaryTypesAtPos' which allows
to set the value of a primary variable by replacing an arbitrary
balance equation. In this case, the index of the primary variable
is Indices::pnIdx, while the index of the replaced equation is
Indices::contiWEqIdx. This allows to set the outflow condition for
equation index Indices::contiNEqIdx.

This completes FS#270.

Reviewed by Timo.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@15167 2fb0f335-1f38-0410-981e-8018bf24f1b0