[multidomain] update references, fix parallel test execution
References have to be updated due to differences between test system and
the system of the reference solution creator. The differences are small,
but still too large for fuzzycompare. This is a known issue, see FS#281
for details.

The two ni-multidomain tests use the same names for the output. This
leads to conflicts when the tests are executed in parallel. The names of
one test are changed to fix this.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@15189 2fb0f335-1f38-0410-981e-8018bf24f1b0