[restart] fix restart functionality for an adaptive grid from dune-alugrid
This solves FS#237, see further details there.

The solution is to be aware of a restart already in the GridCreator.
Then, it is possible to construct the grid there that will be used for
the restarted simulation. This requires that the parameter Problem.Name
is set in the input file in order to deduct the correct restart file
names.

Reviewed and improved by Timo.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@15336 2fb0f335-1f38-0410-981e-8018bf24f1b0