[multidomain] decrease maximum time step size for testing
The tests 2cstokes2p2c and 2cnistokes2p2cni failed due to differences in
the final solutions that were caused by different timesteps that were
caused by a small difference in the number of Newton iterations for one
time step. This patch restricts the maximum time step size from 360 to
240 seconds. By this, the tests succeed. However, this is only a
temporary fix and the basic issue should be investigated as part of our
testing research.

Discussed with fetzer.



git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@15539 2fb0f335-1f38-0410-981e-8018bf24f1b0