[elementgeometry] Add sub control volume geometry member
This is only implemented for 2d cubes for box at the moment.
Saving a geometry might be a nicer option than our internal
geometry calculations. But it has to be investigated what is
faster. The geometries are useful for checkInside operations
for the point source implementation. They can be theoretically
used to calculate scv volumes and scvf normals.