[el2p] Introduced input files, increased simulation time
Tests now use input files instead of command line input.
Simulation time changed to the original setup with 1e6 s of initialisation and 1e6 s of injection. Reference files are adapted accordingly.
Name of the executable for both the parallel and the non-parallel test is now the same.
(reviewed by fetzer)