[range-for] adapt remaining intersection loops in LocalResiduals
In the LocalResiduals, there are several protected functions taking
an IntersectionIterator,
    void eval...(const IntersectionIterator &isIt,

In order to be callable with an (address of an) intersection, the
explicit requirement on the type IntersectionIterator has been
relaxed:
    template<class InetrsectionIterator>
    void eval...(const IntersectionIterator &isIt,

In a future steps these functions will be deprecated and replaced by
equivalents that take intersection objects.

Discussed with and approved by Christoph.