[bboxtree] Implement intersection detection for entities
As Dune currently offers no general solution to find if a
point lies inside an entity, we need an own solution here.
This implements algorithms from different sources (see comments)
for fast detection of whether a point lies in an entity.
Algorithms are only implemented for simplices in all
combinations of dimgrid and dimworld. Other geometries are
subdivided in simplices and then checked for intersection.