[parameters] Don't check consistent type for parameter macro
The check was turned on for debug purposes and exited with an error
if the same parameter was read into object of different type. This
is not an error. For reading vectors it is even an important feature
that it's possible to read it into different types of containers
depending on the usage.

Reviewed by martins