[pointsource] Map from element and scv to pointsource
The mapping from dofIdx to pointsource was inconvenient since the
source functions in dumux are called for each element for each
scv. Now the map maps from a elementIdx scvIdx pair to the source
belonging to the scv. DofSource is thus not the right name anymore
and is now called scvPointSources.