[doxygen] Add features of the sequential tests
A new table for the sequential tests' features was added.
Using <div style="width:70%;overflow:auto;"> limits the horizontal
extent of the html page. As the tables are very broad,
they now have a horizontal scrollbar. This improves the appearance of
the documentation on our weppage.
Maybe there is a smarter idea for this?