[fickslaw] Get difussion coefficients from vol vars
The diffusion coefficient computation consists of three steps.
A scalar diffusion coefficient can be constant or computed depending
on primary and secondary variables in a constitutive law. This is
done for every volume variable once and thus in the volvars update.
In a second step an effective diffusion coefficient in computed
depending on saturation and porosity. In a third step (currently
not implemented) a fibre structure could lead to anisotropic
diffusion tensors. They could be provided in the current framework
as normalized tensors per SCV by the spatial params and then
scaled with the computed diffusion coefficient (see e.g. Linniger 2012).