[mpfa] fix sequential MPFA tests
Sequential MPFA relies on the local vertex numberings of the individual
grid managers. This numberings have changed for the 2d-ALUGrid. Until this
is fixed, change the corresponding tests to use UGGrid instead. Alternatively,
master-old of dune-alugrid can be used.

Also revert changes to the input file that never were reflected in the
corresponding spatial parameters file.