[multidomain, cornerpoint] Update to Dune 2.4, Opm 2015.10, and GCC 6
The installexternal script is updated to fetch and build the correct Opm
modules.

No patch is required anymore for dune-multidomaingrid. In order to keep that
situation, add `#include <numeric>` to the problem files of the multidomain
tests. This is required if the tests are built with GCC 6.

Adapt README files.