[implicit] Remove the shouldWriteOutput check inside writeOutput
When calling writeOutput one would expect that the output is actually
written to file. Instead this might not have happened in the old version.
This commit makes the implicit problem consistent with the sequential
models. Before calling writeOutput one should ask if shouldWriteOutput
is true. This is already consistently done all over dumux, so this
commit removes the double checking.