[pointsource] Add embeddings/setEmbeddings function
The number of embeddings of a point source (number of sub
control volumes the point source belongs to) can now be
set from outside. PointSourceHelper doesn't need to be a
friend class of PointSource anymore. The default number
of embeddings is 1 (a point source with no embeddings, e.g.
outside the grid boundary will be discarded anyway).