[pointsource] Make PointSourceHelper a property
The class computing in which sub control volume a point source falls
gets a property. Thus it can be exchanged with custom implementations
e.g. not making use of the bounding box tree. The default implementation
is the BoundingBoxTreePointSourceHelper using the axis-aligned bounding
box volume hierachy to find in which scv a point source falls.