[bboxtree] Add method to return bbox coordinate pointer
A general method for all dimensions. Call the helper
methods with the actual coordinates pointer instead of the
whole coordinate vector. Technically this is an interface
change. However, the bounding box helper methods only
make sense to be called by the bboxtree itself so it's
an internal class that is necessary for template specialization.