[bboxtree] Add method to intersection two AABB trees
We can collide two AABB trees efficiently now. Thus,
AABB trees of two grids can be intersected. This returns
a pair of vectors. They have the same length. Each entry
contains the element index of the one grid in the first
vector and the intersecting element of the second grid in
the second vector.

This implements primitive tests for intersection of two AABB trees
only for 3d with 1d grids for now. If you want to intersect
grids of other dimensions implement the method collides in the
geometry collision class added with this commit.