[bboxtree] Improve geometry collision and return intersection corners
The geoemtry collision noe return a vector of intersection corners.
The collision algorithm with another budning box tree returns
a vector of BoundingBoxTreeIntersection object.

The object has a method first and second to get the intersecting
entity indices as well as a method corners. From that you
could build e.g. a intersection geometry. The intersections are
not unique yet. There might be interesctions with the same
corner points but different elements.